chore: move unit tests to spatial toggle comp
